Completed hull zone geometry
Builds Polygon based off of the ShipSize and the numbers from front/rear
offsets and conjunctions.  Automatically discards repeat points.  ie the
front arc for the nebulon is a triangle compared to the pentagon of the
cr90.

Tests related to geometry of the nebB pass and match the expected
values.