Edits to Dice Class
Got rid of the valued constructor that gave opportunity to bad dice
generation.  Opted to use a 3 line approach in the calling code instead
of a single line calling a constructor

Modified canHaveFace(DiceFace) to lead to canColorHaveFace since they
shared so much code.  It might still be superflous.

changed "folor" to "color" in a comment

changed 2 methods to camelCase for consistancy.

-----

Updates to DiceRoll to accomodate these changes.