Beginnings of attack selection
Redid turn flow to a single state rather than an overlap of 3 different
Redid flow of update and render based on a switch statement of the turn
flow state
Clickable hullzones to select an attacker, highlight hte defenders, and
select a defender from that list.
Updates to how the yellow dot in the hullzones are handled
Tests to ensure shape intersections works as anticipated
Edits to Range enum to include some hardcoded measuerement values.