Redid how hullzones colors are rendered
Attack selection is depicted on the hullzones themselves, rather than on
an overlay.

This lets the ship graphic show up above the target selection gfx.

Also makes it easier for the attack selection ui to line up with what is
on the screen.  Since we are piggybacking existing geometry and its
rendering, the grunt work is already in place.